swapped page.tsx in homepage with app page
